# auth-service/src/main/java/com/nttdata/authservice/api/dto/LoginRequestDto.java
package com.nttdata.authservice.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequestDto {

    @NotBlank
    private String username;

    @NotBlank
    private String password;
}