# auth-service/src/main/resources/openapi/auth-service-api.yml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: Contract First API for authentication operations
servers:
  - url: http://localhost:7001
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user and issue tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          description: Invalid refresh token
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logout completed
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      operationId: me
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
    TokenRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UserProfileResponse:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string