# auth-service/target/generated-sources/openapi/src/main/java/com/nttdata/authservice/generated/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nttdata.authservice.generated.api;

import com.nttdata.authservice.generated.model.LoginRequest;
import com.nttdata.authservice.generated.model.LoginResponse;
import com.nttdata.authservice.generated.model.LogoutRequest;
import com.nttdata.authservice.generated.model.TokenRefreshRequest;
import com.nttdata.authservice.generated.model.TokenRefreshResponse;
import com.nttdata.authservice.generated.model.UserProfileResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-23T20:37:52.618318600-05:00[America/Lima]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Auth", description = "the Auth API")
public interface AuthApi {

    default AuthApiDelegate getDelegate() {
        return new AuthApiDelegate() {};
    }

    /**
     * POST /api/auth/login : Authenticate user and issue tokens
     *
     * @param loginRequest  (required)
     * @return Authenticated successfully (status code 200)
     *         or Invalid credentials (status code 401)
     */
    @Operation(
        operationId = "login",
        summary = "Authenticate user and issue tokens",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Authenticated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoginResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid credentials")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<LoginResponse>> login(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody Mono<LoginRequest> loginRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().login(loginRequest, exchange);
    }


    /**
     * POST /api/auth/logout : Logout current user
     *
     * @param logoutRequest  (required)
     * @return Logout completed (status code 204)
     */
    @Operation(
        operationId = "logout",
        summary = "Logout current user",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Logout completed")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/logout",
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<Void>> logout(
        @Parameter(name = "LogoutRequest", description = "", required = true) @Valid @RequestBody Mono<LogoutRequest> logoutRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().logout(logoutRequest, exchange);
    }


    /**
     * GET /api/auth/me : Get current authenticated user
     *
     * @return Current user profile (status code 200)
     */
    @Operation(
        operationId = "me",
        summary = "Get current authenticated user",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current user profile", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserProfileResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/auth/me",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<UserProfileResponse>> me(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().me(exchange);
    }


    /**
     * POST /api/auth/refresh : Refresh access token
     *
     * @param tokenRefreshRequest  (required)
     * @return Token refreshed successfully (status code 200)
     *         or Invalid refresh token (status code 401)
     */
    @Operation(
        operationId = "refreshToken",
        summary = "Refresh access token",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Token refreshed successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TokenRefreshResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid refresh token")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/refresh",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<TokenRefreshResponse>> refreshToken(
        @Parameter(name = "TokenRefreshRequest", description = "", required = true) @Valid @RequestBody Mono<TokenRefreshRequest> tokenRefreshRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().refreshToken(tokenRefreshRequest, exchange);
    }

}